Rename "Bin" to "Trash" and "Delete" command.
As discussed in issue [#8863](https://redmine.bbaw.de/issues/8863), labels
related to preliminary deletion are being changed for more intuitive
understanding of the function. "Bin" tabs of the different parts are
henceforth referred to as "Trash" and the command needed to move objects
to those areas are named "Move to Trash".